+++ src/components/PostSummary.js
@@ -37,20 +37,6 @@ function PostSummary(props) {
         }
     }
 
-    function mapTags() {
-        return (
-            <div className="inline-flex w-full justify-end mr-3 text-lg font-extralight">
-                  {props.post.tags.map((tag, i) => 
-                    <div className="mr-2" key={i}>
-                        <Badge pill className="px-2 py-1.5" variant="secondary">
-                            {tag}
-                        </Badge>
-                    </div>
-                    )}     
-            </div>
-        )
-    }
-
     function handleRemove() {
         API.removePost(props.post.id, contextState.roomKey);
     }
@@ -82,26 +68,9 @@ function PostSummary(props) {
                     <br />
                     <span>{props.post.comments.length}</span>
                 </div>
-                <div className="ml-4 mt-2 flex-1 text-left">
-                    <PostHeader post={props.post} setHasSolved={setHasSolved} solved={hasSolved} />
-                </div>
                 <div className="pl-4 pt-2 flex-1 text-left">
-                    <div className="flex justify-between">
-                        {props.post.tags.length > 0 ? mapTags() : null}
-                        {
-                            /* Check if current display name matches name of post. If so allow them to remove it */
-                            contextState.displayName === props.post.author && (
-                                <div className="flex-none pr-4">
-                                    <button
-                                        className="w-8 h-8 flex btn-color rounded-md"
-                                        onClick={handleRemove}
-                                        variant="outline-danger"
-                                    >
-                                        <IoClose className="flex-1 self-center" />
-                                    </button>
-                                </div>
-                            )
-                        }
+                    <div className="flex">
+                        <PostHeader post={props.post} setHasSolved={setHasSolved} solved={hasSolved} />
                     </div>
                     <div className="divide-y">
                         <div className="mx-4 my-2 break-all">{props.post.content}</div>
